jvmcli: Add support for MdST
