+++ build.gradle
@@ -186,7 +186,7 @@ kotlin {
                 implementation 'commons-io:commons-io:2.5'
                 implementation 'org.apache.commons:commons-lang3:3.5'
                 implementation "org.jetbrains.kotlinx:kotlinx-io-jvm:$kotlinio_version"
-                implementation "com.github.cretz.pbandk:pbandk-runtime-jvm:0.3.0"
+                implementation 'com.google.protobuf:protobuf-lite:3.0.1'
                 implementation group: 'xmlpull', name: 'xmlpull', version: '1.1.3.4a'
                 implementation project(':proto')
             }
